* New dependency: `@novivia/babel` * Updated `@novivia/build-module` dependency to v0.5.
* Removed dependencies:
    * `babel-cli`
    * `babel-polyfill`
    * `babel-preset-es2015`
    * `babel-preset-stage-0`
    * `babel-register`
