+++ .eslintrc
@@ -3,6 +3,7 @@
   "root": true,
   "parser": "babel-eslint",
   "env": {
+    "jasmine": true,
     "jest": true,
     "node": true,
     "es6": true
@@ -26,9 +27,9 @@
       "properties": "never"
     }],
     "indent": ["error", 2, {
-      "MemberExpression": 0,
-      "SwitchCase": 1,
-      "VariableDeclarator": 2,
+      "CallExpression": {
+        "arguments": 1
+      },
       "FunctionDeclaration": {
         "body": 1,
         "parameters": 1
@@ -36,7 +37,10 @@
       "FunctionExpression": {
         "body": 1,
         "parameters": 1
-      }
+      },
+      "MemberExpression": 0,
+      "SwitchCase": 1,
+      "VariableDeclarator": 2
     }],
     "key-spacing": "error",
     "keyword-spacing": "error",
@@ -69,7 +73,9 @@
     "no-useless-escape": "error",
     "no-useless-rename": "error",
     "no-whitespace-before-property": "error",
-    "object-property-newline": "error",
+    "object-property-newline": ["error", {
+      "allowMultiplePropertiesPerLine": true
+    }],
     "one-var": ["error", "never"],
     "quotes": "error",
     "rest-spread-spacing": "error",
@@ -149,7 +155,7 @@
       "typeof": true
     }],
     "no-underscore-dangle": ["error", {
-      "allow": ["__", "__DEV__", "__PRERENDER__", "__PROD__", "__STAGE__", "__TEST_ENV__", "_id"],
+      "allow": ["__", "__BUILDING__", "__DEV__", "__PRERENDER__", "__PROD__", "__STAGE__", "__TEST_ENV__", "__TESTING__", "_id"],
       "allowAfterThis": true
     }],
     "no-unexpected-multiline": "error",
@@ -164,14 +170,14 @@
       "args": "after-used"
     }],
     "no-use-before-define": ["error", "nofunc"],
+    "no-useless-return": "error",
     "no-with": "error",
     "unicode-bom": "error",
     "valid-jsdoc": ["error", {
       "matchDescription": "^[A-Z](?:.|\\s)*?[.]$",
-      "prefer": {
-        "returns": "return"
-      },
-      "requireReturn": false
+      "requireParamDescription": false,
+      "requireReturn": false,
+      "requireReturnType": false
     }],
     "valid-typeof": ["error", {
       "requireStringLiterals": true
@@ -243,7 +249,7 @@
     "import/export": "error",
     "import/extensions": ["error", "never"],
     "import/first": "error",
-    "babel/generator-star-spacing": ["error", "after"],
+    "generator-star-spacing": "error",
     "import/max-dependencies": ["error", {
       "max": 30
     }],
@@ -251,7 +257,7 @@
     "babel/new-cap": "error",
     "import/newline-after-import": "error",
     "import/no-absolute-path": "error",
-    "babel/no-await-in-loop": "error",
+    "no-await-in-loop": "error",
     "no-class-assign": "error",
     "no-confusing-arrow": ["error", {
       "allowParens": true
@@ -262,17 +268,19 @@
     "no-dupe-class-members": "error",
     "no-duplicate-imports": "error",
     "no-empty-pattern": "error",
-    "no-invalid-this": 1,
+    "babel/no-invalid-this": "error",
     "no-mixed-requires": "error",
     "import/no-mutable-exports": "error",
-    "no-new-symbol": "error",
+    "import/no-named-default": "error",
     "import/no-namespace": "error",
+    "no-new-symbol": "error",
     "import/no-unresolved": ["error", {
       "amd": true,
       "commonjs": true,
       "ignore": ["build/locales/getLocales$", "js/render/prerenderSimple$"]
     }],
     "no-param-reassign": "error",
+    "no-return-await": "error",
     "no-template-curly-in-string": "error",
     "no-this-before-super": "error",
     "no-useless-concat": "error",
@@ -282,10 +290,10 @@
     "object-shorthand": "error",
     "prefer-arrow-callback": "error",
     "prefer-const": "error",
-    "prefer-reflect": "error",
     "prefer-rest-params": "error",
     "prefer-spread": "error",
     "prefer-template": "error",
+    "require-await": "error",
     "require-yield": "error",
     "strict": ["error", "never"],
     "template-curly-spacing": "error",
@@ -298,11 +306,14 @@
     "flowtype/define-flow-type": "error",
     "flowtype/delimiter-dangle": ["error", "always-multiline"],
     "flowtype/no-dupe-keys": "error",
+    "flowtype/no-primitive-constructor-types": "error",
     "flowtype/object-type-delimiter": "error",
     "flowtype/require-parameter-type": ["error", {
-      "excludeArrowFunctions": "expressionsOnly"
+      "excludeArrowFunctions": true
     }],
-    "flowtype/require-return-type": "error",
+    "flowtype/require-return-type": ["error", {
+      "excludeArrowFunctions": true
+    }],
     "flowtype/require-valid-file-annotation": ["error", "never", {
       "annotationStyle": "block"
     }],
@@ -314,7 +325,7 @@
     "flowtype/space-after-type-colon": "error",
     "flowtype/space-before-generic-bracket": "error",
     "flowtype/space-before-type-colon": "error",
-    "flowtype/type-id-match": "error",
+    "flowtype/type-id-match": ["error", "^([A-Z][A-Za-z0-9]*)+(?:Type)?$"],
     "flowtype/union-intersection-spacing": "error",
     "flowtype/use-flow-type": "error",
     "you-dont-need-lodash-underscore/assign": "error",
@@ -323,7 +334,7 @@
     "you-dont-need-lodash-underscore/fill": "error",
     "you-dont-need-lodash-underscore/find-index": "error",
     "you-dont-need-lodash-underscore/index-of": "error",
-    "you-dont-need-lodash-underscore/is-na-n": "error",
+    "you-dont-need-lodash-underscore/is-nan": "error",
     "you-dont-need-lodash-underscore/join": "error",
     "you-dont-need-lodash-underscore/keys": "error",
     "you-dont-need-lodash-underscore/last-index-of": "error",
@@ -337,6 +348,9 @@
     "block-spacing": ["error", "never"],
     "brace-style": ["error", "1tbs", {
       "allowSingleLine": true
+    }],
+    "capitalized-comments": ["error", "always", {
+      "ignoreConsecutiveComments": true
     }],
     "comma-style": ["error", "last"],
     "computed-property-spacing": ["error", "never"],
@@ -353,8 +367,11 @@
       "before": "never",
       "after": "always"
     }],
-    "max-len": ["error", 80, 2, {
-      "ignoreUrls": true
+    "max-len": ["error", {
+      "code": 80,
+      "ignoreRegExpLiterals": true,
+      "ignoreUrls": true,
+      "tabWidth": 2
     }],
     "newline-per-chained-call": "error",
     "no-floating-decimal": "error",
