+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Versions
 
+## v0.1.2 - (16/12/2016)
+
+* New dependency: `@novivia/babel`
+* Updated `@novivia/build-module` dependency to v0.5.
+
+* Removed dependencies:
+    * `babel-cli`
+    * `babel-polyfill`
+    * `babel-preset-es2015`
+    * `babel-preset-stage-0`
+    * `babel-register`
+
+
 ## v0.1.1 - (02/11/2016)
 
 * Addressed being unable to install on Node 7.x.
