+++ cliWrapper
@@ -1,7 +1,3 @@
 #!/usr/bin/env node
 
-// For dev only.
-// require("babel-register");
-
-require("babel-polyfill");
 require("./cli");
